making authoried requests using  bearer

+++ frontend/src/components/WorkoutDetails.js
@@ -1,20 +1,27 @@
 import { useWorkoutsContext} from '../hooks/useWorkoutsContext';
+import { useAuthContext } from '../hooks/useAuthContext';
 
 import { formatDistanceToNow } from 'date-fns';
 
 const WorkoutDetails = ({workout}) => {
     const { dispatch } = useWorkoutsContext();
+    const { user } = useAuthContext();
 
     const handleClick = async () => {
+        if (!user) {
+            return;
+        }
         const response = await fetch('api/workouts/' + workout._id, {
-            method:'DELETE'
-        })
-        const json = await response.json()
+            method:'DELETE',
+            'Authorisation': `Bearer ${user.token}`
+        });
+        const json = await response.json();
 
         if (response.ok){
-            dispatch({type: 'DELETE_WORKOUT', payload:json})
+            dispatch({type: 'DELETE_WORKOUT', payload:json});
         }
-    }
+    };
+    
     return (
         <div className="workout-details">
             <h4>{workout.title}
